Moved the custom widgets classes to the root gui folder
