--- nw/gui/additions/__init__.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-from nw.gui.additions.qconfiglayout import PagedDialog
-from nw.gui.additions.qconfiglayout import QConfigLayout
-from nw.gui.additions.qconfiglayout import QHelpLabel
-from nw.gui.additions.qconfiglayout import QSwitch
-
-__all__ = [
-    "PagedDialog",
-    "QConfigLayout",
-    "QHelpLabel",
-    "QSwitch",
-]
